requirements.txt
customtkinter
appdirs
Pillow
